requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
vk-api>=11.9.9
mysql-connector-python>=8.2.0
aiohttp>=3.8.3
asyncio>=3.4.3